requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
gunicorn==21.2.0
redis==5.0.1
requests==2.31.0
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
Werkzeug==2.3.7

# GitHub Integration
PyGithub==2.7.0

# Environment Management
python-dotenv==1.0.0

# Data Processing
pandas

# Streamlit
streamlit

# FastAPI (for Render deployment)
fastapi
uvicorn

# Real Ecosystem Integration
supabase>=2.0.0
google-generativeai>=0.3.0
